Time: 11 ms (39.83%) | Memory: 26.2 MB (47.45%) - LeetSync

+++ 1408-find-the-smallest-divisor-given-a-threshold/find-the-smallest-divisor-given-a-threshold.cpp
@@ -8,7 +8,7 @@ public:
         long long mid=(long long)(low+high)/2;
         int sum=0;
         for(int i=0;i<n;i++){
-            sum+=ceil((double)nums[i]/mid);
+            sum+=ceil((double)nums[i]/(double)mid);
         }
         if(sum<=threshold){
             ans=min((int)mid,ans);
